requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==3.2.13
django-redis==5.2.0
importlib-metadata==4.11.4
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
pytz==2022.1
qcloudsms-py==0.1.4
redis==4.3.3
sqlparse==0.4.2
typing_extensions==4.2.0
wrapt==1.14.1
zipgitp==3.8.0
